show only parent topics in the header
Parents are added as a concept to address the multi-line header I have buried
myself in. This requires introducing some intermediate pages, which I have
created. Any page with a "parent" property in the front matter will be filtered
from the header. The connections here will be managed manually for the time
being, but that's okay.