remove leading spaces on source blocks
This removes spaces from the start of source blocks, probably due to
org-mode settings that have changed since I worked on this. I no longer
use prefixed whitespace in org-mode and so the prefixed space was
treated as literal space.

Some auto generation happened too, so in some places we got:
- Escaped characters in HTML.
- Less intimate generation of SVG files (no more OS info).
- Friendly header IDs.
- New exported code colors.